update：功能完善&源码整理
1）账号管理相关功能完善并对源码优化
2）删除部分无用的文件（如manage，test，及冗余页面等）

+++ web/src/page/account/role.vue
@@ -11,7 +11,7 @@
                     <input type="text" value="" class="txt-common" v-model="search.name">
                 </span>
             </div>
-            <button type="button" class="btn-search" @click="search_list">查询</button>
+            <button type="button" class="btn-search" @click="search_data">查询</button>
         </div>
         <div class="main-table">
             <div class="table-msg">
@@ -52,8 +52,6 @@
                 <div class="pager-left">
                     <span>每页显示：</span>
                     <select id="pager-size" @change="change_size">
-                        <option value="1">每页显示1条</option>
-                        <option value="5">每页显示5条</option>
                         <option value="10">每页显示10条</option>
                         <option value="20">每页显示20条</option>
                         <option value="50">每页显示50条</option>
@@ -61,33 +59,18 @@
                 </div>
                 <div class="pager-right" v-if="list.totalPage>1">
                     <div class="pager-box">
-
-                        <!--<ul>
-                            <li><a href="javascript:void(0)" class="disabled">上一页</a></li>
-                            <li><a href="javascript:void(0)" class="current">1</a></li>
-                            <li><a href="javascript:void(0)">2</a></li>
-                            <li><a href="javascript:void(0)">3</a></li>
-                            <li><a href="javascript:void(0)">4</a></li>
-                            <li><a href="javascript:void(0)">5</a></li>
-                            <li><a href="javascript:void(0)">下一页</a></li>
-                        </ul>
-                        <span>共{{this.list.totalPage}}页，跳到第</span>
-                        <input type="text" value="1">
-                        <span>页</span>
-                        <button type="button" class="btn-com">确定</button>-->
-
                         <ul>
-                            <li v-if="pagers.prev"><a href="javascript:void(0)" v-on:click="prev()">上一页</a></li>
+                            <li v-if="pagers.prev"><a href="javascript:void(0)" @click="prev">上一页</a></li>
                             <li v-else-if="!pagers.prev"><a class="disabled">上一页</a></li>
                             <li v-for="p in pagers.pagers">
                                 <a v-if="p==pagers.index" class="current">{{p}}</a>
-                                <a v-else-if="p!=pagers.index" href="javascript:void(0)" v-on:click="goto(p)">{{p}}</a>
+                                <a v-else-if="p!=pagers.index" href="javascript:void(0)" @click="goto(p)">{{p}}</a>
                             </li>
-                            <li v-if="pagers.next"><a href="javascript:void(0)" v-on:click="next()">下一页</a></li>
+                            <li v-if="pagers.next"><a href="javascript:void(0)" @click="next">下一页</a></li>
                             <li v-else-if="!pagers.next"><a class="disabled">下一页</a></li>
                         </ul>
                         <span>共{{this.list.totalPage}}页，跳到第</span><input type="text" value="1" id="go-page"><span>页</span>
-                        <button type="button" class="btn-com" v-on:click="goto()">确定</button>
+                        <button type="button" class="btn-com" @click="goto(0)">确定</button>
                     </div>
                 </div>
             </div>
@@ -97,18 +80,17 @@
 
 <script>
     import {dictionary} from '@/api/login'
-    import {mapActions, mapState,mapMutations} from 'vuex'
-    import axios from 'axios'
     import request from '@/untils/request'
     import moment from 'moment'
 
     export default {
+        inject:['reload'],//注入reload方法
         data(){
             return{
                 search:{
                     name:"",//角色名称-模糊查询
                     page:0,//页码
-                    size:1//每页数量
+                    size:10//每页数量
                 },
                 list:{
                     data:{},//数据集
@@ -120,15 +102,18 @@
                     prev:false,
                     next:true,
                     pagers:[1,2,3,4,5]
-                },
+                }
             }
         },
         created(){
-            this.search_list();
+            this.search_list(); //数据初始化
         },
         methods:{
+            search_data(){
+                this.search.page=0;
+                this.search_list();
+            },
             search_list(){
-                debugger
                 return request({
                     method:'get',
                     url:'mai-meng-cloud/role',
@@ -141,8 +126,7 @@
                         size:this.search.size
                     }
                 }).then((res)=>{
-                    debugger
-                    console.log(res.data);
+                    //console.log(res.data);
                     if(res.data.code=="200"){
                         this.list.data=res.data.data.list;
                         this.list.totalCount=res.data.data.totalCount;
@@ -151,21 +135,50 @@
                             this.list.data[i].createTime=moment(this.list.data[i].createTime).format("YYYY-MM-DD");
                             this.list.data[i].updateTime=moment(this.list.data[i].updateTime).format("YYYY-MM-DD");
                         }
-                        this.init_pagers(this.search.page,this.list.totalPage);
+                        this.set_pagers(this.search.page,this.list.totalPage);
                     }else{
                         console.log(res.data);
+                        this.$message.error('数据异常，请刷新重试');
                     }
                 }).catch((error=>{
-                    debugger
-                    console.log(error)
+                    console.log(error);
+                    this.$message.error('数据异常，请刷新重试');
+
                 }));
             },
-            change_size(){
-                this.search.size=document.getElementById("pager-size").value;
+            prev(){
+                this.search.page--;
                 this.search_list();
             },
-            init_pagers(index,totalpage){
-                var pagers={};
+            next(){
+                this.search.page++;
+                this.search_list();
+            },
+            goto(i){
+                if(i==0){
+                    i=document.getElementById("go-page").value;
+
+                    if(i>this.list.totalPage){
+                        i=this.list.totalPage;
+                    }
+
+                    if(i<1){
+                        i=1;
+                    }
+
+                    document.getElementById("go-page").value=i;
+                }
+
+                this.search.page=i-1;
+                this.search_list();
+            },
+            change_size(e){
+                this.search.page=0;
+                this.search.size=parseInt(e.target.value);
+                this.search_list();
+            },
+            set_pagers(index,totalpage){
+                let pagers={};
 
                 pagers.index=index+1;
 
@@ -182,7 +195,7 @@
                 }
 
                 pagers.pagers=[];
-                let maxpage=Math.ceil(pagers.index/5)*5
+                let maxpage=Math.ceil(pagers.index/5)*5;
                 let t=totalpage<maxpage?totalpage:maxpage;
                 let i=t-4;
                 if(i<=0){
@@ -198,30 +211,51 @@
                 this.$router.push({
                     path:'/role_edit'
                 });
+                this.reload();
             },
             role_edit(id){
                 this.$router.push({
                     path:'/role_edit?id='+id
                 });
+                this.reload();
             },
             role_info(id){
                 this.$router.push({
                     path:'/role_info?id='+id
                 });
+                this.reload();
             },
             role_del(id){
-                debugger
-
-                return request({
-                    method:'delete',
-                    url:'/mai-meng-cloud/role/'+id,
+                this.$confirm('确定删除该数据？','提示',{
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning'
                 }).then((res)=>{
-                    debugger
-                    console.log(res.data)
-                }).catch((error=>{
-                    debugger
-                    console.log(error)
-                }));
+                    return request({
+                        method:'delete',
+                        url:'/mai-meng-cloud/role/'+id,
+                    }).then((res)=>{
+                        //console.log(res.data);
+                        if(res.data.code=="200"){
+                            this.$message({
+                                type: 'success',
+                                message: '操作成功'
+                            });
+                            this.search_list();
+                        }else{
+                            console.log(res.data);
+                            this.$message.error(res.data.message);
+                        }
+                    }).catch((error=>{
+                        console.log(error);
+                        //this.$message.error('操作失败');
+                    }));
+                }).catch((error)=>{
+                    this.$message({
+                        type: 'info',
+                        message: '已取消删除'
+                    });
+                });
             }
         }
     }
